New release

+++ install/getcontent.sh
@@ -14,10 +14,23 @@ if [ ! -d "$DIR" ]; then
 	mkdir $DIR
 fi
 
-curl $URL/$FNAME -o $DIR/$FNAME
+http_code=$(curl --speed-limit 5 --speed-time 30 --write-out '\n%{http_code}\n' $URL/$FNAME -o $DIR/$FNAME | tail -n 1)
+if [ $? -ne 0 ]; then
+	exit 1
+fi
+
+if [ $http_code -ne 200 ]; then
+	echo "***" \[$http_code\] "Server Error ***"
+	exit 1
+fi
 
 tar -tf $DIR/$FNAME
+if [ $? -ne 0 ]; then
+	exit 1
+fi
 
 tar -xzvf $DIR/$FNAME -C ../
 
 rm $DIR/$FNAME
+
+exit 0
